Make interaction with the message a side effect
The library currently requires that message handlers return a value
in order to react to the message. This is not ideal for two reasons:

1)	Handlers cannot do multiple things with the message,
	such as replying twice, reacting and then replying, etc.
2)	Handlers cannot reply immediately and then continue processing.

I have changed the API from returning a Reply object to passing the args
for the Reply constructor to message.reply(). Additionally, I have made
marking the message as read optional, as some bots may not want to send
read receipts, or may only want to send them when the response is done.

Additionally, I have cleaned up some logging statements.
Messages that are really only useful for debugging were moved from INFO
to DEBUG, error logs were moved to ERROR,
and tracebacks were added using the `exc_info` kwarg when necessary.